+++ app/src/main/cpp/planner/include/QuadraticCalculator.h
@@ -13,10 +13,11 @@ namespace nav_core
     class QuadraticCalculator : public PotentialCalculator
     {
     public:
-        QuadraticCalculator(int nx, int ny): PotentialCalculator(nx,ny) {}
+        QuadraticCalculator(unsigned int nx, unsigned int ny) : PotentialCalculator(nx, ny)
+        {}
 
-        float calculatePotential(float* potential,
-                                 unsigned char cost, int n, float prev_potential);
+        double calculatePotential(double *potential,
+                                  unsigned char cost, int n, double prev_potential);
     };
 } //end namespace nav_core
